# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ComentarioDTO": {
				"properties": {
					"idComentario": {
						"type": "number",
						"format": "double"
					},
					"idUsuario": {
						"type": "number",
						"format": "double"
					},
					"idProducto": {
						"type": "number",
						"format": "double"
					},
					"comentario": {
						"type": "string"
					},
					"fechaComentario": {
						"type": "string",
						"format": "date-time"
					},
					"estado": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"idComentario",
					"idUsuario",
					"idProducto",
					"comentario",
					"fechaComentario",
					"estado",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CrearComentarioBody": {
				"properties": {
					"idUsuario": {
						"type": "number",
						"format": "double"
					},
					"idProducto": {
						"type": "number",
						"format": "double"
					},
					"comentario": {
						"type": "string"
					}
				},
				"required": [
					"idUsuario",
					"idProducto",
					"comentario"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ActualizarComentarioBody": {
				"properties": {
					"comentario": {
						"type": "string"
					}
				},
				"required": [
					"comentario"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"legacyJwt": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "producttrack",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "@mariaximena0724 y emilia"
		}
	},
	"paths": {
...
		"/comentarios/{productoId}": {
			"get": {
				"operationId": "GetComentariosPorProducto",
				"responses": {
					"200": {
						"description": "Lista de comentarios.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ComentarioDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Obtiene los comentarios asociados a un producto.\r\n\r\n👉 [Documentación completa del controlador en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/comentarios.controller)",
				"tags": [
					"Comentarios"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID del producto.",
						"in": "path",
						"name": "productoId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/comentarios": {
			"post": {
				"operationId": "CrearComentario",
				"responses": {
					"200": {
						"description": "Comentario creado.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ComentarioDTO"
								}
							}
						}
					}
				},
				"description": "Crea un nuevo comentario para un producto.\r\n\r\n👉 [Documentación completa del controlador en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/comentarios.controller)",
				"tags": [
					"Comentarios"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Datos del comentario.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CrearComentarioBody",
								"description": "Datos del comentario."
							}
						}
					}
				}
			}
		},
		"/comentarios/{idComentario}": {
			"put": {
				"operationId": "ActualizarComentario",
				"responses": {
					"200": {
						"description": "Comentario actualizado.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ComentarioDTO"
								}
							}
						}
					}
				},
				"description": "Actualiza un comentario existente.\r\n\r\n👉 [Documentación completa del controlador en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/comentarios.controller)",
				"tags": [
					"Comentarios"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID del comentario.",
						"in": "path",
						"name": "idComentario",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Nuevo contenido del comentario.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ActualizarComentarioBody",
								"description": "Nuevo contenido del comentario."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "EliminarComentario",
				"responses": {
					"200": {
						"description": "Mensaje de confirmación.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"mensaje": {
											"type": "string"
										}
									},
									"required": [
										"mensaje"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Elimina un comentario existente.\r\n\r\n👉 [Documentación completa del controlador en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/comentarios.controller)",
				"tags": [
					"Comentarios"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID del comentario.",
						"in": "path",
						"name": "idComentario",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}